server/pyproject.toml
[tool.poetry]
name = "server"
version = "0.1.0"
description = ""
authors = ["Catminusminus <resonafolma@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
foolbox = "^2.3.0"
fastapi = "^0.42.0"
tensorflow = "1.14"
keras = "^2.3.1"
Pillow = "^6.2.1"
uvicorn = "^0.10.4"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
